+++ world.c
@@ -338,7 +338,7 @@ void SV_LinkEdict (prvm_edict_t *ent, qboolean touch_triggers)
 
 	if (ent->fields.server->solid == SOLID_BSP)
 	{
-		int modelindex = ent->fields.server->modelindex;
+		int modelindex = (int)ent->fields.server->modelindex;
 		if (modelindex < 0 || modelindex > MAX_MODELS)
 		{
 			Con_Printf("edict %i: SOLID_BSP with invalid modelindex!\n", PRVM_NUM_FOR_EDICT(ent));
@@ -442,7 +442,7 @@ trace_t SV_ClipMoveToEntity(prvm_edict_t *ent, const vec3_t start, const vec3_t 
 
 	if ((int) ent->fields.server->solid == SOLID_BSP || movetype == MOVE_HITMODEL)
 	{
-		unsigned int modelindex = ent->fields.server->modelindex;
+		unsigned int modelindex = (unsigned int)ent->fields.server->modelindex;
 		// if the modelindex is 0, it shouldn't be SOLID_BSP!
 		if (modelindex == 0)
 		{
